Fix remaining golangci-lint issues
- Fix gci import ordering in memdiag.go
- Fix intrange: use Go 1.22+ integer range syntax
- Fix prealloc: pre-allocate slices where size is known
- Fix unconvert: remove unnecessary conversion
- Fix wrapcheck: wrap context.Err() and file operation errors
- Fix zerologlint: add nolint comment for intentional Event() pattern
- Add nolint comments for intentional complexity:
  - gocognit/gocyclo/cyclop/funlen/maintidx for complex functions
  - nestif for deeply nested test validation
  - revive for intentional type name stuttering (InventoryRow/Reader)
  - unparam for future-proofing error return
- Name return values in runMergeBuildPhase for gocritic

+++ pkg/sysmem/mem_linux.go
@@ -11,5 +11,5 @@ func totalSystemMemory() (uint64, bool) {
 		return 0, false
 	}
 	// Total RAM in bytes = Totalram * Unit
-	return uint64(info.Totalram) * uint64(info.Unit), true
+	return info.Totalram * uint64(info.Unit), true
 }
